@@ -1,6 +1,6 @@
 {
   "name": "@nerdwallet/shepherd",
-  "version": "2.6.1",
+  "version": "3.0.0",
   "description": "A utility for applying code changes across many repositories",
   "type": "module",
   "keywords": [
@@ -37,7 +37,7 @@
     "lint": "eslint src/**/*.ts && prettier --check .",
     "prepublishOnly": "yarn test && yarn build",
     "test": "jest --coverage src/",
-    "e2e-test": "npm run build && jest --config=jest.e2e.config.cjs e2e/"
+    "e2e-test": "npm run build && jest --config=jest.e2e.config.js e2e/"
   },
   "files": [
     "lib"
@@ -50,6 +50,7 @@
     "@types/js-yaml": "^4.0.9",
     "chalk": "^5.3.0",
     "child-process-promise": "^2.2.1",
+    "table": "^6.8.1",
     "commander": "^12.1.0",
     "fs-extra": "^11.2.0",
     "joi": "^17.13.3",
@@ -76,18 +77,18 @@
     "@typescript-eslint/eslint-plugin": "^8.8.0",
     "@typescript-eslint/parser": "^8.8.0",
     "babel-jest": "^29.7.0",
-    "conventional-changelog-conventionalcommits": "^8.0.0",
+    "conventional-changelog-conventionalcommits": "^9.0.0",
     "eslint": "^9.12.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-config-prettier": "^10.0.0",
+    "eslint-import-resolver-typescript": "^4.0.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-prefer-arrow": "^1.2.3",
-    "globals": "^15.10.0",
+    "globals": "^16.0.0",
     "jest": "^29.7.0",
     "prettier": "^3.3.3",
     "semantic-release": "^24.1.2",
     "ts-jest": "^29.2.5",
-    "typescript": "5.7",
+    "typescript": "5.9",
     "typescript-eslint": "^8.8.0"
   },
   "publishConfig": {
