@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   singleQuote: true,
   trailingComma: 'es5',
   printWidth: 100,
