@@ -290,7 +290,7 @@ class GithubAdapter extends GitAdapter {
   }
 
   public getBaseBranch(repo: IRepo): string {
-    return repo.defaultBranch;
+    return process.env.SHEPHERD_BASE_BRANCH ?? repo.defaultBranch;
   }
 
   public async getEnvironmentVariables(repo: IRepo): Promise<IEnvironmentVariables> {
@@ -316,6 +316,43 @@ class GithubAdapter extends GitAdapter {
       throw new Error(`Unknown protocol ${githubProtocol}. Valid values are 'ssh' and 'https'`);
     }
   }
+
+  public createIssue = async (repo: IRepo): Promise<string> => {
+    const {
+      migration: { spec },
+    } = this.migrationContext;
+    const { owner, name } = repo;
+    const { issues } = spec;
+    //Create an issue with the title , issue message and labels
+
+    return await this.githubService.createAndGetIssueNumber({
+      owner,
+      repo: name,
+      title: issues?.title || 0,
+      body: issues?.description,
+      labels: issues?.labels,
+    });
+  };
+
+  public updateIssue = async (repo: IRepo, issueNumber: number): Promise<void> => {
+    const {
+      migration: { spec },
+    } = this.migrationContext;
+    const { owner, name } = repo;
+    const { issues } = spec;
+
+    //Update an issue's title , issue message and labels
+    await this.githubService.updateIssue({
+      owner,
+      repo: name,
+      title: issues?.title?.toString(),
+      issue_number: issueNumber,
+      body: issues?.description,
+      labels: issues?.labels,
+      state: issues?.state,
+      state_reason: issues?.state_reason,
+    });
+  };
 
   private async checkActionSafety(repo: IRepo): Promise<SafetyStatus> {
     const { owner, name } = repo;
