@@ -1,7 +1,7 @@
-import list from './list';
-import { IMigrationContext } from '../migration-context';
-import mockAdapter from '../adapters/adapter.mock';
-import mockLogger from '../logger/logger.mock';
+import list from './list.js';
+import { IMigrationContext } from '../migration-context.js';
+import mockAdapter from '../adapters/adapter.mock.js';
+import mockLogger from '../logger/logger.mock.js';
 
 describe('list commmand', () => {
   let mockContext: IMigrationContext;
