@@ -1,18 +1,22 @@
 import { IMigrationContext } from '../migration-context.js';
 import forEachRepo from '../util/for-each-repo.js';
 
-export default async (context: IMigrationContext) => {
+const determineRepoPRStatus = async (context: IMigrationContext, repo: any) => {
   const { logger, adapter } = context;
+  const spinner = logger.spinner('Determining repo PR status');
 
+  try {
+    const status = await adapter.getPullRequestStatus(repo);
+    spinner.destroy();
+    status.forEach((s) => logger.info(s));
+  } catch (error: any) {
+    logger.error(error);
+    spinner.fail('Failed to determine PR status');
+  }
+};
+
+export default async (context: IMigrationContext) => {
   await forEachRepo(context, async (repo) => {
-    const spinner = logger.spinner('Determining repo PR status');
-    try {
-      const status = await adapter.getPullRequestStatus(repo);
-      spinner.destroy();
-      status.forEach((s) => logger.info(s));
-    } catch (e: any) {
-      logger.error(e);
-      spinner.fail('Failed to determine PR status');
-    }
+    await determineRepoPRStatus(context, repo);
   });
 };
